Be consistent about which preprocessor define you use to enable/disable the use of ReedSolomon across all of the codebase.
(Though, granted, the p25 RS code is extremely useful. The DMR RS code suffers from the designers being permanently retarded when they chose the polynomial size and the truncation offset. So we're in GF(255) but can only correct 4 errors because what DMR calls RS(12, 9, 4) is a truncation of, and I quote, RS(255, 252, 4) code. (Also they actually mean RS(255, 251, 4) but whatever, close enough). Problem is it does have forward-correcting capability, it's not just a checksum, so there is marginal utility in calling it).

Also, fuck the DMR spec for the following line: "NOTE: polynomial coefficients are in hexadecimal radix, while exponents are in decimal radix".

Best part is the exponents range from 0-3, so you could just say both are in hex and still be correct, but with less confusion along the way!